modify Reservation class

+++ src/main/java/FlyAway/reservation/dto/CreateReservationDto.java
@@ -3,13 +3,14 @@ package FlyAway.reservation.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreateReservationDto (
         @NotNull
-        Long price,
+        BigDecimal price,
         @NotNull
-        Long seatNumber,
+        Integer seatNumber,
         @NotNull
         Long clientId,
         @NotNull
