+++ src/main/java/FlyAway/reservation/dto/DisplayReservationDto.java
@@ -1,13 +1,14 @@
 package FlyAway.reservation.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record DisplayReservationDto(
         UUID id,
         LocalDateTime reservationDate,
-        Long price,
-        Long seatNumber,
+        BigDecimal price,
+        Integer seatNumber,
         Boolean cancelled,
         Long clientId,
         UUID flightId
