+++ src/main/java/FlyAway/reservation/dto/ReservationDto.java
@@ -3,13 +3,14 @@ package FlyAway.reservation.dto;
 import FlyAway.client.dto.ClientDto;
 import FlyAway.flight.dto.FlightDto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 
 public record ReservationDto(
         LocalDateTime reservationDate,
-        Long price,
-        Long seatNumber,
+        BigDecimal price,
+        Integer seatNumber,
         Boolean cancelled,
         ClientDto clientDto,
         FlightDto flightDto
