+++ src/main/java/FlyAway/reservation/dto/ReservationWithoutUserDto.java
@@ -2,12 +2,13 @@ package FlyAway.reservation.dto;
 
 import FlyAway.flight.dto.FlightDto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ReservationWithoutUserDto(
         LocalDateTime reservationDate,
-        Long price,
-        Long seatNumber,
+        BigDecimal price,
+        Integer seatNumber,
         Boolean cancelled,
         FlightDto flightDto
 ) {
